moved placeholder contracts to data
